@@ -2,7 +2,7 @@
   "name": "@giftdibs/ux",
   "version": "1.0.0-rc.1",
   "peerDependencies": {
-    "@angular/common": "^7.0.0",
-    "@angular/core": "^7.0.0"
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0"
   }
 }
